[MERGEDOG] Use MPS-supported dtype in convolution test
Failing job: macos-py3-arm64 / test (mps, 1, 1, macos-m2-15)
Failing test: nn/test_convolution.py::TestConvolutionNNDeviceTypeMPS::test_slow_conv_transpose3d_kernel_size_mismatch_mps
CI repro command:
```
PYTORCH_TEST_WITH_SLOW=1 python test/nn/test_convolution.py TestConvolutionNNDeviceTypeMPS.test_slow_conv_transpose3d_kernel_size_mismatch_mps
```

Error:
```
TypeError: Cannot convert a MPS Tensor to float64 dtype as the MPS framework doesn't support float64.
```

The device-generic regression test created its input, weight, and bias as
float64 tensors before reaching the shape validation under test. Use the
default float32 dtype, which MPS supports and which does not affect the
kernel-size mismatch validation.

Test Plan:
```
git diff --check
```

The CI test was not run locally, as required by the mergedog environment.

Authored with an AI assistant.

+++ test/nn/test_convolution.py
@@ -1182,8 +1182,8 @@ class TestConvolutionNN(NNTestCase):
 
 class TestConvolutionNNDeviceType(NNTestCase):
     def test_slow_conv_transpose3d_kernel_size_mismatch(self, device):
-        inp = torch.full((1, 2, 4, 5, 4), 0.5, dtype=torch.double, device=device)
-        weight = torch.full((2, 3, 2, 3, 2), 0.5, dtype=torch.double, device=device)
+        inp = torch.full((1, 2, 4, 5, 4), 0.5, device=device)
+        weight = torch.full((2, 3, 2, 3, 2), 0.5, device=device)
         with self.assertRaisesRegex(
             RuntimeError, "kernel_size.*must match weight spatial dimensions"
         ):
@@ -1191,7 +1191,7 @@ class TestConvolutionNNDeviceType(NNTestCase):
                 inp,
                 weight,
                 [1, 1, 1],
-                torch.full((3,), 0.5, dtype=torch.double, device=device),
+                torch.full((3,), 0.5, device=device),
                 [1, 1, 1],
                 [2, 2, 2],
                 [0, 0, 0],
